Parallelize stable table sizes monitoring script
