--- sql/monitoring/telemetry_stable_sizes_v1/metadata.yaml
@@ -1,7 +1,0 @@
-friendly_name: Telemetry Stable Sizes
-description: >
-  Table sizes of telemetry_stable tables, partitioned by day.
-owners:
-  - ascholtz@mozilla.com
-labels:
-  schedule: daily
